#15 move extension files to extension folder
