+++ extension/cypress/plugins/index.js
