+++ extension/cypress/set-up-server.js
@@ -1,9 +1,9 @@
 const url = require('url')
 
-const serverFn = require('../extension/src/server')
-const httpUtils = require('../extension/src/utils')
-const ctpClientBuilder = require('../extension/src/ctp/ctp-client')
-const { routes } = require('../extension/src/routes')
+const serverFn = require('../src/server')
+const httpUtils = require('../src/utils')
+const ctpClientBuilder = require('../src/ctp/ctp-client')
+const { routes } = require('../src/routes')
 
 async function addTestRoutes () {
   /**
