+++ extension/cypress/setUp/credit_card_3d_redirect.js
@@ -6,8 +6,8 @@ global.navigator = {}
 
 const adyenEncrypt = require('adyen-cse-js')
 const _ = require('lodash')
-const ctpClientBuilder = require('../../extension/src/ctp/ctp-client')
-const payment3dTemplate = require('../../extension/test/fixtures/payment-credit-card-3d.json')
+const ctpClientBuilder = require('../../src/ctp/ctp-client')
+const payment3dTemplate = require('../../test/fixtures/payment-credit-card-3d.json')
 
 const key = process.env.CLIENT_ENCRYPTION_PUBLIC_KEY
