+++ extension/package.json
@@ -3,10 +3,9 @@
   "version": "0.0.1",
   "description": "Integration between Commercetools and Adyen payment service provider",
   "scripts": {
-    "test": "mocha --exit --timeout 30000 --full-trace extension/test/**/*.spec.js",
+    "test": "mocha --exit --timeout 30000 --full-trace test/**/*.spec.js",
     "extension": "node ./extensions/src/init.js",
-    "lint": "eslint extension",
-    "init": "CTP_PROJECT_KEY=projectKey CTP_CLIENT_ID=clientId CTP_CLIENT_SECRET=clientSecret API_EXTENSION_BASE_URL=extensionUrl node ./extension/scripts/initResources.js",
+    "lint": "eslint .",
     "cypress:run": "cypress run",
     "cypress:open": "cypress open",
     "cypress": "server-test cypress-setup 8000 cypress:run",
@@ -43,8 +42,6 @@
     "chai": "^4.2.0",
     "cypress": "^3.1.5",
     "eslint": "^5.12.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-commercetools": "^6.0.0",
     "eslint-plugin-cypress": "^2.2.0",
     "eslint-plugin-import": "^2.14.0",
     "mocha": "^5.2.0",
