+++ extension/src/api/health/health.controller.js
@@ -1,5 +1,6 @@
 const httpUtils = require('../../utils')
 
+// todo: implement proper health
 function processRequest (request, response) {
   httpUtils.sendResponse(response, 200)
 }
